refactor(assert): split collection equality helper
